fix no.5

+++ asafov.andrew/P4/main.cpp
@@ -4,7 +4,7 @@
 
 int main()
 {
-  size_t lenght = 0;
+  unsigned long long lenght = 0;
   char* str = asafov::inputString(lenght);
   if(lenght == 0)
   {
@@ -14,7 +14,7 @@ int main()
   int count = asafov::countUniqLetters(str, lenght);
   std::cout<<count<<'\n';
   char* unusedLetters = asafov::getUnusedLetters(str, lenght);
-  for(size_t i = 0; unusedLetters[i] != 0; i++)
+  for(unsigned long long i = 0; unusedLetters[i] != 0; i++)
   {
     std::cout<<unusedLetters[i];
   }
