+++ asafov.andrew/P4/resize.cpp
@@ -3,7 +3,7 @@
 char* asafov::resize(char* string, const unsigned long long lenght) {
   char* new_string = new char[lenght + 8];
   char* old_string = string;
-  for (size_t i = 0; i < lenght; i++) {
+  for (unsigned long long i = 0; i < lenght; i++) {
     new_string[i] = old_string[i];
   }
   delete[] old_string;
